fix(scripts): Skip lock file generation for workspaces (#17398)

+++ scripts/publish-starters.sh
@@ -14,6 +14,7 @@ for folder in $GLOB; do
   cd $BASE
 
   NAME=$(cat $folder/package.json | jq -r '.name')
+  IS_WORKSPACE=$(cat $folder/package.json | jq -r '.workspaces')
   CLONE_DIR="__${NAME}__clone__"
   
   # sync to read-only clones
@@ -25,8 +26,10 @@ for folder in $GLOB; do
   find . | grep -v ".git" | grep -v "^\.*$" | xargs rm -rf # delete all files (to handle deletions in monorepo)
   cp -r $BASE/$folder/. .
 
-  rm -rf yarn.lock
-  yarn import # generate a new yarn.lock file based on package-lock.json
+  if [ "$IS_WORKSPACE" = null ]; then
+    rm -rf yarn.lock
+    yarn import # generate a new yarn.lock file based on package-lock.json
+  fi
 
   git add .
   git commit --message "$COMMIT_MESSAGE"
